+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Unreleased Changes
 
 ### Breaking Changes
+- It is no longer possible to set both `--checkpoint-sync-url` and `--initial-state`. If your node fails to start after upgrade, ensure that only one of these is set.
 
 ### Additions and Improvements
 - Updated Teku bootnode ENR on Sepolia and Mainnet
@@ -14,3 +15,4 @@
 - Scheduled Electra for Gnosis at epoch 1337856 (April 30, 2025, 14:03:40 UTC)
 
 ### Bug Fixes
+ - It is no longer possible to set both `--checkpoint-sync-url` and `--initial-state`.